chore(federation): split some operation functionality into modules (#5336)
